Fixed meal_page_flip and some other bugs
Changed meal_page_flip in LyfeMainWindow.py so that clicking on the Meal button properly takes you to the Meal window.
Made multiple small changes to MealControllerClass.py, including commenting out code that currently causes errors due to not yet being implemented, changing variables that will need to be numerical data types to ints (may need to be floats, depending on how precise we want to be), and deleted unnecessary screen change at the end of add_the_meal.