Changed process of adding meal to database
Now meals will be inserted into the database based on the date and time that the meal was eaten. The more recent meal will be closer to the top of the file, and the older meals will be towards the bottom. FindInsertIndex takes a meal's date and time data and figures out where to place the meal based on that.