Multiple changes
List of changes:
Added text under the required fields (time and date if entered incorrectly, name and calories if empty) that's only shown when conditions are met. Changes field border to red to indicate wrong fields.
Added a function that checks that the inputs are correct before adding the information into the database.
Added a function to make a new window back to the main meal UI to be called in multiple areas.
Added a back button to go back to the main meal UI if the user doesn't want to enter a meal.